move aside C sources that have a (faster) asm counterpart
this was neccessary because some ancient commit removed the explicit
listing of source files in the Makefile.
The original intention was to keep the C sources for reference and
in case we'd want to port to another platform.
This makes sure (again) that the faster version is picked up for the library.