Add include of optional Make.config.local to Makefile
The intention is to provide a way to play around with the configuration variables such as ONLY_68K etc. and CFLAGS without modifying Makefile and maybe forgetting to undo such changes. Make.config.local is now added to .gitignore to ensure that these local configuration will not be pushed to the repository.