Remove the access mode macros from stdio.h
These macros are already defined in Mintlib's fcntl.h and should be included from there.

The removed values were the ones Pure C uses and some of them are different. That can easily lead to unexpected issues when using library functions like open() which receive these flags in a parameter.